requirements.txt
aiofiles==23.2.1
aiogram==3.15.0
aiohttp==3.10.11
aiosignal==1.3.1
APScheduler==3.10.4
attrs==24.2.0
beautifulsoup4==4.12.3
bs4==0.0.1
cachetools==5.3.3
certifi==2024.8.30
charset-normalizer==3.4.1
flyerapi==1.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
oauthlib==3.2.2
peewee==3.17.9
propcache==0.2.0
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.6.4
pydantic_core==2.16.3
pyee==12.1.1
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==21.10
pytz==2024.1
redis==5.0.1
requests==2.32.3
rsa==4.9
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.3.0
yarl==1.17.0